refactor app dependencies and app to diffrent files file
